Inphi: Use periodic timer to resolve 25G lane recovery
Some of the PLLs of retimer are not getting locked even after the PHY
init is completed. This keeps the PHY in an unusable state until
In112525S03PhyStartup function is invoked with ifconfig command on the
interface.

To fix this issue UEFI periodic timer is used to check for gLaneRecovery
every 5 seconds to reinitiate In112525S03LaneRecovery if required.

As In112525S03LaneRecovery is called periodically, gExitBootServicesEvent
for lane recovery is removed.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Calvin Johnson <calvin.johnson@nxp.com>