Platform/SolidRun: Initial LX2160aCex7 commit
This is the initial platform commit for the SolidRun CEX7 based
platforms based on the NXP LX2160a SOC.  Many of the firmware specific
bits are derived from the LX2160aRdb platform so this commit is based
on the existing support and customized for implementation specific
differences.

ACPI based changes will be integrated in the next commit.

Signed-off-by: Jon Nettleton <jon@solid-run.com>