Platform/SolidRun: Enable networking for LX2160aCex7 platform
This patch enables networking support for the LX2160Cex7 platform
with a minimal memory allocation of 512MB's of memory for the firmware.
This of course can be modified back to the default 2GB's of memory
that is allocated for high end networking appliances like the LX2160ARDB.

Prebuilt binaries can be loaded from edk2-non-osi, or built from source
and the paths in the fdf changed accordingly.

Additionally the default fdsize may need to be altered in order to
compensate for larger firmware images in the future.

Signed-off-by: Jon Nettleton <jon@solid-run.com>