Silicon/NXP: Pause the McBin firmware before loading the rich OS.
We put the MCBin firmware in Pause mode after finalizing the
configuration that will be used by the rich OS.  This allows the
SMMU to be configured properly to talk to the device.  It is
up to the drivers in the rich OS to then unpause the firmware
before using it.

Signed-off-by: Jon Nettleton <jon@solid-run.com>