Silicon/NXP: Fixup driver memory allocations to work with declared Hobs
Because the memory allocations are happening during the virtual memory mapping
the memory space can not just be allocated but instead needs to be looked up
and remapped appropriately.  Change the drivers that rely on this to first
try and find the Hob that defines the memory space and only if that fails do
we try and allocate the memory address space required by the driver.

Signed-off-by: Jon Nettleton <jon@solid-run.com>