Silicon/NXP: LX2160a CEX7 fix PWM fan cooling
This commit fixes the PWM Fan control for the CEX7 module.  It
will be further expanded to allow the trip points and fan speed
to be customized in the gui, but for now these settings should
be a good starting point for functional fan control.  The default
configuration is setup in this method.

1) We use the TMU temperature as the control trigger for the
different fan states.  The external PWM thermal monitor is not
that useful until we add Auto fan mode support in.
2) Add GPIO control for the Fan Full speed pin.  This is pulled
high by the RCW so if the board sits in the firmware or crashes
on boot the system will have maximum cooling.  By using this
as its own FAN device it also allows us to have more flexibility
with regards to the FAN Low and Fan High speeds that are used
in a different thermal zone.
3) For Fan0 and Fan1 we add an _INI method that configures the
PWM for software DCY mode.  When this was in the main _INI for
the TMZ device it was not getting set causing the PWM changes
to not take effect.
4) Because we now have essentially 4 cooling states I have
setup this base functionality.

0-50C - 33% Duty
50-65C - 50% Duty
65-80C - 80% Duty
> 80C - 100% Duty

Signed-off-by: Jon Nettleton <jon@solid-run.com>