Silicon/NXP: LX2160aCEX7 DSDT fixes for errata
This patch accomplishes 3 things.  The first is to
remove the FSPI DSDT definitions as that is used by
UEFI as either a boot medium or variable storage and
therefore should not be exposed to the host OS.

Secondly we address an errata with the LX2160a silicon.
3.3V I/O can cause damage to the SOC over extended
periods of time.  To help stop this from happening we
are limiting the the uSDHC interface to only support
UHS modes in the host os.

Lastly we are seeing issues with DDR timings on the V2
silicon.  Therefore we limit the eMMC to HS200 speeds
at 200MHz.

Signed-off-by: Jon Nettleton <jon@solid-run.com>