nxp: lx2160acex7: Add PcdDriverConvertsFvbFuncPointers to True
It was found that the VariableRuntime was unconditionally
converting the FVB function pointers for runtime, which was
causing a double conversion and putting our runtime pointers
into a non-executable memory region. In order to not break
existing codebases I have introduced PcdDriverConvertsFvbFuncPointers
to edk2 and we set it to true to avoid the double pointer
conversion for runtime.

Signed-off-by: Jon Nettleton <jon@solid-run.com>