NXP: Pcf2129Rtc start RTC on invalid time bit
Currently we are checking for the invalid data bit and returning
a default hwclock date and time.  However unlock the standard Linux
drivers most SBSA systems expect the clock to be running even though
the date is invalid.  Therefore in the case where the data is invalid
we call LibSetTime to set the time and start the RTC running

Signed-off-by: Jon Nettleton <jon@solid-run.com>