SysEeprom: Add TLV support
This patch adds support to parsing TLV header that is flashed by default
on SolidRun LX2160A COM express modules.

Notice that MacNo starts with the value '1' and the wrap around makes it
possible that the RJ45 connector dpmac17 is actually the first MAC
address mentioned in the TLV MAC field.

Signed-off-by: Rabeeh Khoury <rabeeh@solid-run.com>