Silicon/NXP: Move reserved memory for mc-bus to start of high mem
Ram sizes can change but are generally all above the first low
memory window.  We use high memory for the packet processor but
instead of putting it at the end of the high memory address range
we put it at the beginning.  This means that we have a constant
base address and only need to account for PcdDpaa2McHighRamSize.

This makes accounting for the range much simpler and makes it
trivial to account for this range when building our RMR tables.

Signed-off-by: Jon Nettleton <jon@solid-run.com>