Silicon/NXP: LX2160a rework UpdateDsdtPcie to UpdateDsdt
Moving forward there will be more runtime DSDT changes we will
need to make.  This formwat is a bit more readable, and since
it won't just be for PCIe we make the function more generic.

This also adds a change to the PCIe configs setting _BBN to 1
rather than zero as both ESXi and BSDs have requested this.

Signed-off-by: Jon Nettleton <jon@solid-run.com>