Silicon/NXP: LX2160a CEX7 A few table fixups
The first is most relevant and that is fixing the duplicate UUID
in the Tmu asl descriptions.

The second less relevant but added for consistency in preparation
to moving things upstream is renaming the USB nodes to XHC and also
switching the _CID to _HID and removing the legacy _HID.  This
is the descriptions being used by other edk2-platform descriptions
for XHCI nodes.

Signed-off-by: Jon Nettleton <jon@solid-run.com>